requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydub>=0.25.0
faster-whisper>=1.0.0
